Move DSORenderer to its own files
